Changed PS1 to fix overlapping printout
The terminal prompt set via the PS1 variable to include name of current git branch (if current dir is a git repo or subfolder of git repo) previously had problems sometimes with printed text overlapping the prompt. 
For example if you do the "reverse-i search" the printout from that would overlap the prompt in a ugly way and then when picking a result from the reverse-i search the printout would be all messed up and you could not see where the cursor was positioned properly.

This commit fixes these issues for the prompt and at least so far I haven't found any problems with it. It also seems like wrapping text output when the text surpasses the width of the terminal window is also handled better than previous to this commit.